pyproject.toml
[project]
name = "repl-nix-buzzscriptgenerator-1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "trafilatura>=1.12.2",
    "google-generativeai>=0.8.3",
    "beautifulsoup4>=4.12.3",
    "requests>=2.32.3",
    "youtube-transcript-api>=0.6.2",
    "webdriver-manager>=4.0.2",
    "selenium>=4.26.1",
    "urllib3",
    "werkzeug",
]
